### 2019-05-04 #1429 @whedon
[PRE REVIEW]: phylosmith: an R-package for reproducible and efficient microbiome analysis with phyloseq-objects
**Submitting author:** @schuyler-smith (<a href="http://orcid.org/0000-0001-5720-5611">Schuyler Smith</a>)
**Repository:** <a href="https://github.com/schuyler-smith/phylosmith/" target ="_blank">https://github.com/schuyler-smith/phylosmith/</a>
**Version:** 0.99.0
