### 2019-05-04 #1429 @kyleniemeyer
Hi @schuyler-smith, your paper file needs to be named `paper.md` and `paper.bib` for our system to detect it in your repo; could you fix that?
