### 2019-05-07 c8adba5
update figure themes, render site, add sp to DESCRIPTION
