+++ R/graphs.R
@@ -102,25 +102,29 @@ abundance_heatmap_ggplot <- function(phyloseq_obj, classification = NULL,
     setkey(graph_data, 'Sample', 'Abundance')
     set(graph_data, which(graph_data[['Abundance']] == 0), 'Abundance', NA)
 
-    g <- ggplot(graph_data, aes_string('Sample', classification,
-        fill = 'Abundance')) +
+    g <- ggplot(graph_data, aes_string('Sample', classification, fill = 'Abundance')) +
         geom_tile(color = "white", size = 0.25) +
-        facet_grid(treatment_name, scales = "free", space = "free") +
-        theme_classic() +
-        theme(
-          axis.text.x = element_text(angle = -35, hjust = 0, size = 12),
-          axis.text.y = element_text(hjust = 0.95, size = 12),
-          axis.title.x=element_blank(),
-          axis.title.y=element_blank(),
-          legend.title=element_blank(),
-          legend.text=element_text(size = 16),
-          legend.spacing.x = unit(0.2, 'cm'),
-          legend.background = element_rect(fill = (alpha = 0))
-        ) +
-        if(transformation %in% c('none', 'relative_abundance')){
-            scale_fill_gradientn(colors = graph_colors)
-        } else {scale_fill_gradientn(colors = graph_colors,
-            trans = transformation)}
+        facet_grid(treatment_name, scales = "free", space = "free")
+    g <- g + theme_classic() +
+      theme(
+        axis.text.x = element_text(angle = -30, hjust = 0, size = 12),
+        axis.text.y = element_text(hjust = 0.95, size = 12),
+        axis.title.x=element_blank(),
+        axis.title.y=element_blank(),
+        axis.ticks.x = element_blank(),
+        legend.title=element_blank(),
+        legend.text=element_text(size = 16),
+        legend.spacing.x = unit(0.2, 'cm'),
+        legend.background = element_rect(fill = (alpha = 0), color = 'black', size = 0.25),
+        panel.background = element_rect(color = 'black', size = 1.4),
+        strip.text.x = element_text(size = 12, face = 'bold'),
+        strip.background = element_rect(colour = 'black', size = 1.4)
+      ) +
+      scale_x_discrete(expand = expand_scale(mult = 0, add = .53)) +
+      if(transformation %in% c('none', 'relative_abundance')){
+          scale_fill_gradientn(colors = graph_colors)
+      } else {scale_fill_gradientn(colors = graph_colors,
+          trans = transformation)}
     return(g)
 }
 
@@ -156,8 +160,7 @@ abundance_heatmap_ggplot <- function(phyloseq_obj, classification = NULL,
 #' @import ggplot2
 #' @export
 #' @return ggplot-object
-#' @examples abundance_lines_ggplot(soil_column, classification = 'phylum',
-#' treatment = c('Matrix', 'Treatment'), relative_abundance = TRUE)
+#' @examples abundance_lines_ggplot(soil_column, classification = 'phylum', treatment = c('Matrix', 'Treatment'), relative_abundance = TRUE)
 
 abundance_lines_ggplot <- function(phyloseq_obj, classification = NULL,
     treatment, subset = NULL, relative_abundance = FALSE, points = TRUE,
@@ -205,7 +208,7 @@ abundance_lines_ggplot <- function(phyloseq_obj, classification = NULL,
     if(relative_abundance){phyloseq_obj <- relative_abundance(phyloseq_obj)}
     treatment_name <- paste(treatment, collapse = sep)
 
-    if(!(is.null(classification))){
+    if(is.null(classification)){
         classification <- 'OTU'
         graph_data <- phyloseq(
             access(phyloseq_obj, 'otu_table'),
[…]
