+++ R/graphs.R
@@ -686,7 +686,7 @@ phylogeny_profile_ggplot <- function(phyloseq_obj, classification = NULL,
       g <- g + facet_grid(treatment_name, scales = "free", space = "free")
     }
     if(merge){
-      g <- g + geom_bar(aes_string(color = classification, fill = classification), stat = 'identity', position = 'stack', size = 0.2, width = 0.98)
+      g <- g + geom_bar(aes_string(color = classification, fill = classification), stat = 'identity', position = 'stack', size = 0.2, width = 0.95)
     } else {
       g <- g + geom_bar(stat = "identity", position = "stack", size = 0.12, color = 'black')}
     if(relative_abundance){g <- g + ylab('Relative Abundance')}
