### 2019-05-08 8f7e124
replace heatmap default with viridis
