### 2019-05-08 be6a637
add layout function, change network layout
