+++ R/graphs.R
@@ -105,11 +105,11 @@ abundance_heatmap_ggplot <- function(phyloseq_obj, classification = NULL,
       theme(
         axis.text.x = element_text(angle = 30, hjust = 1, size = 12),
         axis.text.y = element_text(hjust = 0.95, size = 12),
-        axis.title.x=element_blank(),
-        axis.title.y=element_blank(),
+        axis.title.x = element_blank(),
+        axis.title.y = element_blank(),
         axis.ticks.x = element_blank(),
-        legend.title=element_blank(),
-        legend.text=element_text(size = 16),
+        legend.title = element_text(size = 16, face = 'bold'),
+        legend.text = element_text(size = 16),
         legend.spacing.x = unit(0.2, 'cm'),
         legend.background = element_rect(fill = (alpha = 0), color = 'black', size = 0.25),
         panel.background = element_rect(color = 'black', size = 1.4),
@@ -127,11 +127,11 @@ abundance_heatmap_ggplot <- function(phyloseq_obj, classification = NULL,
         }
       } else {
         if(colors == 'default'){
-          scale_fill_viridis(trans = transformation)
+          scale_fill_viridis(trans = transformation, name = transformation)
         } else {
           color_count <- 100
           graph_colors <- create_palette(color_count, colors)
-          scale_fill_gradientn(colors = graph_colors, trans = transformation)
+          scale_fill_gradientn(colors = graph_colors, trans = transformation, name = transformation)
         }
       }
     return(g)
@@ -272,14 +272,14 @@ abundance_lines_ggplot <- function(phyloseq_obj, classification = NULL,
     return(g)
 }
 
-#' Create an object of the abundance barplots from a phyloseq object.
+#' Create a ggplot object of the co-occurrence from a phyloseq object.
 #' Function from the phylosmith-package.
 #'
 #' This function takes a \code{\link[phyloseq]{phyloseq-class}} object and
-#' creates barplots of taxa by treatment.
+#' creates a network from the co-occurrence.
 #' @useDynLib phylosmith
 #' @usage network_phyloseq(phyloseq_obj, classification = NULL,
-#' treatment = NULL, subset = NULL, co_occurrence_table = NULL,
+#' treatment = NULL, subset = NULL, co_occurrence_table = NULL, layout = NULL,
 #' nodes_of_interest = NULL, node_colors = 'default',
 #' cluster = FALSE, cluster_colors = 'default', buffer = 0.5)
 #' @param phyloseq_obj A \code{\link[phyloseq]{phyloseq-class}} object. It
@@ -300,6 +300,8 @@ abundance_lines_ggplot <- function(phyloseq_obj, classification = NULL,
 #' \code{phyloseq_obj}, computed using \code{\link{co_occurrence}}. If no
 #' table is given, it will be computed with the \code{phyloseq_obj}, using the
 #' given \code{treatment} and \code{p} = 0.05.
+#' @param layout (optional) an igraph layout of the network, for reproducibility.
+#' Can be created with \code{\link{graph_layout_phyloseq}}.
 #' @param nodes_of_interest A vector of names of classes within the
 #' \code{classification} to be labeled.
 #' @param node_colors Name of a color set from the
@@ -324,11 +326,11 @@ abundance_lines_ggplot <- function(phyloseq_obj, classification = NULL,
 #' @return ggplot-object
 #' @examples
 #' #network_phyloseq(soil_column, treatment = c('Matrix', 'Treatment'),
-#' #subset = 'Soil_Manure', co_occurrence_table = NULL,
[…]
