+++ man/graph_layout_phyloseq.Rd
@@ -1,0 +1,47 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/graphs.R
+\name{graph_layout_phyloseq}
+\alias{graph_layout_phyloseq}
+\title{Create an igraph_layout object of the co-occurrence from a phyloseq object.
+Function from the phylosmith-package.}
+\usage{
+graph_layout_phyloseq(phyloseq_obj, classification = NULL,
+treatment = NULL, subset = NULL, co_occurrence_table = NULL,
+algorithm = 'fr')
+}
+\arguments{
+\item{phyloseq_obj}{A \code{\link[phyloseq]{phyloseq-class}} object. It
+must contain \code{\link[phyloseq:sample_data]{sample_data()}}) with
+information about each sample, and it must contain
+\code{\link[phyloseq:tax_table]{tax_table()}}) with information about each
+taxa/gene.}
+
+\item{classification}{Column name as a string or number in the
+\code{\link[phyloseq:tax_table]{tax_table}} for the factor to use for node
+colors.}
+
+\item{treatment}{Column name as a string or number in the
+\code{\link[phyloseq:sample_data]{sample_data}}. This can be a vector of
+multiple columns and they will be combined into a new column.}
+
+\item{subset}{A factor within the \code{treatment}. This will remove any
+samples that to not contain this factor. This can be a vector of multiple
+factors to subset on.}
+
+\item{co_occurrence_table}{Table of the co-occurrence of taxa/genes in the
+\code{phyloseq_obj}, computed using \code{\link{co_occurrence}}. If no
+table is given, it will be computed with the \code{phyloseq_obj}, using the
+given \code{treatment} and \code{p} = 0.05.}
+
+\item{algorithm}{Supported \code{\link[igraph:layout_]{igraph::layout_()}} algorithm.}
+}
+\value{
+igraph_layout object
+}
+\description{
+Create an igraph_layout object of the co-occurrence from a phyloseq object.
+}
+\examples{
+#graph_layout_phyloseq(soil_column, treatment = c('Matrix', 'Treatment'),
+#subset = 'Soil_Manure', co_occurrence_table = NULL, algorithm = 'fr')
+}
