+++ NAMESPACE
@@ -6,6 +6,7 @@ export(co_occurrence)
 export(common_taxa)
 export(conglomerate_samples)
 export(conglomerate_taxa)
+export(graph_layout_phyloseq)
 export(histogram_permuted_rhos)
 export(melt_phyloseq)
 export(merge_treatments)
