### 2019-05-09 #1429 @schuyler-smith
> Hi @schuyler-smith, your paper file needs to be named `paper.md` and `paper.bib` for our system to detect it in your repo; could you fix that?

@karthik

Sorry for the late response! For some reason the notifications of the messages were not pushing to me. Here I was wondering why I hadn't heard from anyone, and it was on my end!

I renamed the files. Was the fail because of the travis build? I have openmp in my c++ code and so I had messed up my configure file for the flags when dealing with native MacOS that doesn't support it. It should be fixed, and able to compile on all systems now!
