### 2019-05-09 #1429 @schuyler-smith
@kyleniemeyer @karthik
Thank you! I was able to generate the paper. Please let me know if I missed anything else, I tried my best to be diligent in the submission checklist.
