### 2019-05-09 a6fa079
rename network functions, comomn taxa for dataset, update docs
