+++ R/graphs.R
@@ -1,8 +1,10 @@
-#' Create a ggplot object of the heatmap of the abundance table from a
+#' Create a heatmap ggplot object of the abundance table from a
 #' phyloseq object. Function from the phylosmith-package.
 #'
-#' This function takes a \code{\link[phyloseq]{phyloseq-class}} object and
-#' creates heatmaps of the abundances across samples.
+#' Inputs a \code{\link[phyloseq]{phyloseq-class}} object and
+#' creates heatmaps of the abundances across samples as a ggplot object.
+#' The default color choice is the viridis palette, which is supposed to
+#' be both aesthetic for normal and color-blind viewers.
 #' @useDynLib phylosmith
 #' @usage abundance_heatmap_ggplot(phyloseq_obj, classification = NULL,
 #' treatment, subset = NULL,
@@ -137,11 +139,11 @@ abundance_heatmap_ggplot <- function(phyloseq_obj, classification = NULL,
     return(g)
 }
 
-#' Create a ggplot object of the abundance table from a phyloseq object.
+#' Create a lineplot ggplot object of the abundance table from a phyloseq object.
 #' Function from the phylosmith-package.
 #'
-#' This function takes a \code{\link[phyloseq]{phyloseq-class}} object and
-#' creates line graphs across samples.
+#' Inputs a \code{\link[phyloseq]{phyloseq-class}} object and
+#' creates line graphs with points across samples.
 #' @useDynLib phylosmith
 #' @usage abundance_lines_ggplot(phyloseq_obj, classification = NULL,
 #' treatment, subset = NULL,
@@ -169,7 +171,8 @@ abundance_heatmap_ggplot <- function(phyloseq_obj, classification = NULL,
 #' @import ggplot2
 #' @export
 #' @return ggplot-object
-#' @examples abundance_lines_ggplot(soil_column, classification = 'phylum', treatment = c('Matrix', 'Treatment'), relative_abundance = TRUE)
+#' @examples abundance_lines_ggplot(soil_column, classification = 'phylum',
+#' treatment = c('Matrix', 'Treatment'), relative_abundance = TRUE)
 
 abundance_lines_ggplot <- function(phyloseq_obj, classification = NULL,
     treatment, subset = NULL, relative_abundance = FALSE, points = TRUE,
@@ -272,13 +275,15 @@ abundance_lines_ggplot <- function(phyloseq_obj, classification = NULL,
     return(g)
 }
 
-#' Create a ggplot object of the co-occurrence from a phyloseq object.
+#' Create a node network ggplot object of the co-occurrence from a phyloseq object.
 #' Function from the phylosmith-package.
 #'
-#' This function takes a \code{\link[phyloseq]{phyloseq-class}} object and
-#' creates a network from the co-occurrence.
+#' Inputs a \code{\link[phyloseq]{phyloseq-class}} object and
+#' creates a network from the co-occurrence. The co-occurrence can either be
+#' input, or it will be calculated with the Spearman-rank correlation. Also,
+#' the layout of the graph can be given as an argument as well for reprodusibility.
 #' @useDynLib phylosmith
-#' @usage network_phyloseq(phyloseq_obj, classification = NULL,
+#' @usage network_ps(phyloseq_obj, classification = NULL,
 #' treatment = NULL, subset = NULL, co_occurrence_table = NULL, layout = NULL,
 #' nodes_of_interest = NULL, node_colors = 'default',
 #' cluster = FALSE, cluster_colors = 'default', buffer = 0.5)
@@ -301,7 +306,7 @@ abundance_lines_ggplot <- function(phyloseq_obj, classification = NULL,
[…]
