+++ R/co_occurrence.R
@@ -292,7 +292,8 @@ quantile_permuted_rhos <- function(permuted_rhos, p = 0.05,
 #' @export
 #' @return ggplot
 #' @examples
-#' permuted_rhos <- permute_rho(soil_column, treatment = c('Matrix', 'Treatment'), replicate_samples = 'Day', permutations = 1,  cores = 0)
+#' permuted_rhos <- permute_rho(soil_column, treatment = c('Matrix', 'Treatment'),
+#' replicate_samples = 'Day', permutations = 1,  cores = 0)
 #' histogram_permuted_rhos(permuted_rhos)
 #' histogram_permuted_rhos(permuted_rhos, p = 0.01)
 
