### 2019-05-10 84895dc
add dois to bib, add JOSS badge
