+++ JOSS_paper/paper.bib
@@ -43,10 +43,18 @@
 }
 
 @misc{carey,
-  title={Ten simple rules for biologists learning to program},
-  author={Carey, Maureen A and Papin, Jason A},
-  year={2018},
-  publisher={Public Library of Science}
+  author = {Carey, Maureen A. AND Papin, Jason A.},
+  journal = {PLOS Computational Biology},
+  publisher = {Public Library of Science},
+  title = {Ten simple rules for biologists learning to program},
+  year = {2018},
+  month = {01},
+  volume = {14},
+  url = {https://doi.org/10.1371/journal.pcbi.1005871},
+  pages = {1-11},
+  abstract = {},
+  number = {1},
+  doi = {10.1371/journal.pcbi.1005871}
 }
 
 @article{tsne,
@@ -67,7 +75,8 @@
   number={7402},
   pages={207},
   year={2012},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi = {10.1038/nature11234}
 }
 
 @article{generic2,
