### 2019-05-13 5c8cacd
fix taxa filter factor order keeping method
