### 2019-05-13 fab8ff2
move treatment name out of if treatment taxa filter
