### 2019-05-21 03202c3
remove network example.. runs but errors.. unsure
