+++ R/graphs.R
@@ -330,9 +330,9 @@ abundance_lines_ggplot <- function(phyloseq_obj, classification = NULL,
 #' @export
 #' @return ggplot-object
 #' @examples
-#' network_ps(soil_column, treatment = c('Matrix', 'Treatment'),
-#' subset = 'Soil Manure', co_occurrence_table = NULL, layout = NULL,
-#' classification = 'phylum')
+#' #network_ps(soil_column, treatment = c('Matrix', 'Treatment'),
+#' #subset = 'Soil Manure', co_occurrence_table = NULL, layout = NULL,
+#' #classification = 'phylum')
 
 network_ps <- function(phyloseq_obj, classification = NULL,
     treatment = NULL, subset = NULL, co_occurrence_table = NULL,
