+++ DESCRIPTION
@@ -26,6 +26,7 @@ Imports:
     igraph,
     ggraph,
     ggrepel,
+    dplyr,
     sf,
     sp,
     vegan,
