+++ .travis.yml
@@ -19,6 +19,7 @@ addons:
    packages:
      - libudunits2-dev # required for units, required for ggforce, required for ggraph
      - libgdal-dev # required for sf
+     - libgeos-dev 
 env:
   global:
   - R_REMOTES_NO_ERRORS_FROM_WARNINGS=true
@@ -28,12 +29,4 @@ env:
   - TRAVIS_CXXFLAGS="-Wall -Wextra -pedantic -Werror"
 
 r_binary_packages:
-  - dplyr
-  - sp
   - ggplot2
-
-# before_install:
-#   - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then brew install llvm &&
-#     export PATH="/usr/local/opt/llvm/bin:$PATH" &&
-#     export LDFLAGS="-L/usr/local/opt/llvm/lib" &&
-#     export CFLAGS="-I/usr/local/opt/llvm/include"; fi
