### 2019-05-22 21fcab0
copy sf_dep travis lines
