+++ .travis.yml
@@ -14,13 +14,10 @@ os:
 compiler:
   - gcc
   - clang
-addons:
- apt:
-   packages:
-     - libudunits2-dev # required for units, required for ggforce, required for ggraph
-     - libgdal-dev # required for sf
-     - libgeos-dev 
-     - libproj-dev
+before_install:
+  - sudo add-apt-repository ppa:ubuntugis/ppa --yes
+  - sudo apt-get --yes --force-yes update -qq
+  - sudo apt-get install --yes libudunits2-dev libproj-dev libgeos-dev libgdal-dev
 env:
   global:
   - R_REMOTES_NO_ERRORS_FROM_WARNINGS=true
@@ -29,8 +26,6 @@ env:
   # remove linux
   - TRAVIS_CXXFLAGS="-Wall -Wextra -pedantic -Werror"
 
-r_packages:
-  - rgdal
-
 r_binary_packages:
   - ggplot2
+
