### 2019-05-23 1ca0ed5
remove trees in taxa conglomerations, fix x-axis in abundance bars
