+++ R/graphs.R
@@ -957,7 +957,7 @@ taxa_abundance_bars_ggplot <- function(phyloseq_obj, classification = NULL,
                                    linetype = 'solid'),
         axis.line.y = element_line(colour = 'black', size = 1,
                                    linetype = 'solid'),
-        axis.text.x = element_text(size = 12, vjust = 0.7, hjust = 1, angle = 30),
+        axis.text.x = element_text(size = 12, vjust = 1, hjust = 1, angle = 30),
         axis.text.y = element_text(size = 12),
         axis.title.x = element_text(size = 16, face = "bold"),
         axis.title.y = element_text(size = 16, face = "bold"),
