### 2019-05-28 a4511f8
fix refseq handling
