+++ man/soil_column.Rd
@@ -6,13 +6,17 @@
 \title{Soil Column 16S Data - OTUs}
 \format{A \code{\link[phyloseq]{phyloseq-class}} object.}
 \source{
-Simulation
+Experiment
 }
 \usage{
 data(soil_column)
 }
 \description{
-Soil Column 16S Data - OTUs
+This dataset was from an experiment that looked at how microbes from
+manure used as fertilizer move through soil systems. The manure was
+applied to the soil in a controlled system, and rain simulations were
+conducted. The soil was tested at various time points, as was the water
+that ran through the soil (effluent).
 }
 \references{
 \url{https://www.frontiersin.org/articles/10.3389/fmicb.2018.03197/full}
