### 2019-05-29 a1a3576
update examples for updatated dataset, add subset to cooccurrence
