+++ R/graphs.R
@@ -110,9 +110,9 @@ abundance_heatmap_ggplot <- function(phyloseq_obj, classification = NULL,
         axis.title.x = element_blank(),
         axis.title.y = element_blank(),
         axis.ticks.x = element_blank(),
-        legend.title = element_text(size = 16, face = 'bold'),
-        legend.text = element_text(size = 16),
-        legend.spacing.x = unit(0.2, 'cm'),
+        legend.title = element_text(size = 12, face = 'bold'),
+        legend.text = element_text(size = 12),
+        legend.spacing.x = unit(0.005, 'npc'),
         legend.background = element_rect(fill = (alpha = 0), color = 'black', size = 0.25),
         panel.background = element_rect(color = 'black', size = 1.4),
         strip.text.x = element_text(size = 12, face = 'bold'),
@@ -259,11 +259,11 @@ abundance_lines_ggplot <- function(phyloseq_obj, classification = NULL,
         axis.text.x = element_text(angle = 30, hjust = 1, size = 12),
         axis.text.y = element_text(hjust = 0.95, size = 12),
         axis.title.x = element_blank(),
-        axis.title.y = element_text(size = 16),
+        axis.title.y = element_text(size = 12),
         axis.ticks.x = element_blank(),
-        legend.title = element_text(size = 16, face = 'bold'),
-        legend.text = element_text(size = 16),
-        legend.spacing.x = unit(0.05, 'cm'),
+        legend.title = element_text(size = 12, face = 'bold'),
+        legend.text = element_text(size = 12),
+        legend.spacing.x = unit(0.005, 'npc'),
         legend.background = element_rect(fill = (alpha = 0)),
         panel.background = element_rect(color = 'black', size = 1.5, fill = 'white'),
         panel.spacing = unit(.015, 'npc'),
@@ -473,8 +473,8 @@ network_ps <- function(phyloseq_obj, classification = NULL,
                                          box.padding = unit(0.8, "lines"), point.padding = unit(0.1, "lines"), size = 5, show.legend = FALSE)
     }
     g <- g + theme(
-      legend.text = element_text(size = 16),
-      legend.title = element_text(size = 16, face = 'bold'),
+      legend.text = element_text(size = 12),
+      legend.title = element_text(size = 12, face = 'bold'),
       legend.spacing.x = unit(0.005, 'npc')
     )
     return(g)
@@ -686,11 +686,11 @@ nmds_phyloseq_ggplot <- function(phyloseq_obj, treatment, circle = TRUE,
                                    linetype = 'solid'),
         axis.text.x = element_text(size = 12),
         axis.text.y = element_text(size = 12),
-        axis.title.x = element_text(size = 16, face= "bold"),
-        axis.title.y = element_text(size = 16, face= "bold"),
-        legend.title = element_text(size = 16, face = 'bold'),
-        legend.text = element_text(size = 16),
-        legend.spacing.x = unit(0.2, 'cm'),
+        axis.title.x = element_text(size = 12, face= "bold"),
+        axis.title.y = element_text(size = 12, face= "bold"),
+        legend.title = element_text(size = 12, face = 'bold'),
+        legend.text = element_text(size = 12),
+        legend.spacing.x = unit(0.005, 'npc'),
         legend.background = element_rect(fill = (alpha = 0))
       ) + labs(x = 'NMDS Dimension 1', y = 'NMDS Dimension 2')
     return(g)
@@ -819,11 +819,11 @@ phylogeny_profile_ggplot <- function(phyloseq_obj, classification = NULL,
         axis.text.x = element_text(angle = 30, hjust = 1, size = 12),
         axis.text.y = element_text(size = 12),
[…]
