+++ R/data_wrangling.R
@@ -28,20 +28,20 @@
 
 common_taxa <- function(phyloseq_obj, treatment = NULL, subset = NULL, n = 'all'){
     if(!inherits(phyloseq_obj, "phyloseq")){
-        stop("common_taxa(): `phyloseq_obj` must be a phyloseq-class object",
+        stop("`phyloseq_obj` must be a phyloseq-class object",
             call. = FALSE)
     }
     if(is.null(access(phyloseq_obj, 'sam_data'))){
-        stop("common_taxa(): `phyloseq_obj` must contain sample_data()
+        stop("`phyloseq_obj` must contain sample_data()
         information", call. = FALSE)
     }
     treatment <- check_numeric_treatment(phyloseq_obj, treatment)
     if(any(!(treatment %in% colnames(access(phyloseq_obj, 'sam_data'))))){
-        stop("common_taxa(): `treatment` must be at least one column name, or
+        stop("`treatment` must be at least one column name, or
         index, from the sample_data()", call. = FALSE)
     }
     if(!(is.numeric(n)) & n != 'all'){
-        stop("common_taxa(): `n` must be either 'all' or a numeric value less
+        stop("`n` must be either 'all' or a numeric value less
         than the number of treatments being compared", call. = FALSE)
     }
     phyloseq_obj <- phyloseq(access(phyloseq_obj, 'otu_table'),
@@ -52,7 +52,7 @@ common_taxa <- function(phyloseq_obj, treatment = NULL, subset = NULL, n = 'all'
                                        'sam_data')[[treatment_name]])
     if(n == 'all'){n <- length(treatment_classes)}
     if(n > length(treatment_classes)){
-      stop("common_taxa(): `n` must be either 'all' or a numeric value less
+      stop("`n` must be either 'all' or a numeric value less
         than the number of treatments being compared", call. = FALSE)
     }
     seen_taxa <- lapply(treatment_classes, FUN = function(trt){
@@ -107,21 +107,21 @@ common_taxa <- function(phyloseq_obj, treatment = NULL, subset = NULL, n = 'all'
 conglomerate_samples <- function(phyloseq_obj, treatment, subset = NULL,
     merge_on = treatment){
     if(!inherits(phyloseq_obj, "phyloseq")){
-        stop("conglomerate_samples(): `phyloseq_obj` must be a phyloseq-class
+        stop("`phyloseq_obj` must be a phyloseq-class
         object", call. = FALSE)
     }
     if(is.null(access(phyloseq_obj, 'sam_data'))){
-        stop("conglomerate_samples(): `phyloseq_obj` must contain
+        stop("`phyloseq_obj` must contain
         sample_data() information", call. = FALSE)
     }
     treatment <- check_numeric_treatment(phyloseq_obj, treatment)
     if(any(!(treatment %in% colnames(access(phyloseq_obj, 'sam_data'))))){
-        stop("conglomerate_samples(): `treatment` must be at least one column
+        stop("`treatment` must be at least one column
         name, or index, from the sample_data()", call. = FALSE)
     }
     merge_on <- check_numeric_treatment(phyloseq_obj, merge_on)
     if(any(!(merge_on %in% colnames(access(phyloseq_obj, 'sam_data'))))){
-        stop("conglomerate_samples(): `merge_on` must be at least one column
+        stop("`merge_on` must be at least one column
         name, or index, from the sample_data()", call. = FALSE)
     }
     if(!(is.null(access(phyloseq_obj, 'phy_tree')))){
@@ -246,21 +246,21 @@ conglomerate_samples <- function(phyloseq_obj, treatment, subset = NULL,
[…]
