+++ man/check_numeric_treatment.Rd
@@ -23,15 +23,3 @@ string
 \description{
 Converts numeric values to column names in sample_data.
 }
-\examples{
-check_numeric_treatment <- function(phyloseq_obj, ...){
-treatments <- list(...)
-treatments <- unlist(sapply(treatments, FUN = function(treatment){
-  if(is.numeric(treatment)){
-    return(colnames(access(phyloseq_obj, 'sam_data')[,treatment]))
-  } else {return(treatment)}
-}))
-return(treatments)
-}
-check_numeric_treatment(soil_column, 2)
-}

[…]
