### 2019-05-31 e319e41
change graph title text sizes
