+++ R/graphs.R
@@ -110,12 +110,12 @@ abundance_heatmap_ggplot <- function(phyloseq_obj, classification = NULL,
         axis.title.x = element_blank(),
         axis.title.y = element_blank(),
         axis.ticks.x = element_blank(),
-        legend.title = element_text(size = 12, face = 'bold'),
+        legend.title = element_text(size = 14, face = 'bold'),
         legend.text = element_text(size = 12),
         legend.spacing.x = unit(0.005, 'npc'),
         legend.background = element_rect(fill = (alpha = 0), color = 'black', size = 0.25),
         panel.background = element_rect(color = 'black', size = 1.4),
-        strip.text.x = element_text(size = 12, face = 'bold'),
+        strip.text.x = element_text(size = 14, face = 'bold'),
         strip.background = element_rect(colour = 'black', size = 1.4)
       ) +
       scale_x_discrete(expand = expand_scale(mult = 0, add = .53)) +
@@ -258,16 +258,16 @@ abundance_lines_ggplot <- function(phyloseq_obj, classification = NULL,
       theme(
         axis.text.x = element_text(angle = 30, hjust = 1, size = 12),
         axis.text.y = element_text(hjust = 0.95, size = 12),
-        axis.title.x = element_blank(),
-        axis.title.y = element_text(size = 12),
+        axis.title.x = element_text(size = 14, face = 'bold'),
+        axis.title.y = element_text(size = 14, face = 'bold'),
         axis.ticks.x = element_blank(),
-        legend.title = element_text(size = 12, face = 'bold'),
+        legend.title = element_text(size = 14, face = 'bold'),
         legend.text = element_text(size = 12),
         legend.spacing.x = unit(0.005, 'npc'),
         legend.background = element_rect(fill = (alpha = 0)),
         panel.background = element_rect(color = 'black', size = 1.5, fill = 'white'),
         panel.spacing = unit(.015, 'npc'),
-        strip.text.x = element_text(size = 12, face = 'bold', color = 'black'),
+        strip.text.x = element_text(size = 14, face = 'bold', color = 'black'),
         strip.background = element_rect(colour = 'black', size = 1.4, fill = 'white')
       ) +
       scale_y_continuous(expand = expand_scale(mult = c(0, 0.003), add = c(0.0015, 0.001))) +
@@ -686,9 +686,9 @@ nmds_phyloseq_ggplot <- function(phyloseq_obj, treatment, circle = TRUE,
                                    linetype = 'solid'),
         axis.text.x = element_text(size = 12),
         axis.text.y = element_text(size = 12),
-        axis.title.x = element_text(size = 12, face= "bold"),
-        axis.title.y = element_text(size = 12, face= "bold"),
-        legend.title = element_text(size = 12, face = 'bold'),
+        axis.title.x = element_text(size = 14, face= "bold"),
+        axis.title.y = element_text(size = 14, face= "bold"),
+        legend.title = element_text(size = 14, face = 'bold'),
         legend.text = element_text(size = 12),
         legend.spacing.x = unit(0.005, 'npc'),
         legend.background = element_rect(fill = (alpha = 0))
@@ -818,15 +818,15 @@ phylogeny_profile_ggplot <- function(phyloseq_obj, classification = NULL,
       theme(
         axis.text.x = element_text(angle = 30, hjust = 1, size = 12),
         axis.text.y = element_text(size = 12),
-        axis.title.x = element_blank(),
-        axis.title.y = element_text(size = 12, face = 'bold'),
+        axis.title.x = element_text(size = 14, face = 'bold'),
+        axis.title.y = element_text(size = 14, face = 'bold'),
         axis.ticks.x = element_blank(),
[…]
