### 2019-06-03 e1ea21e
update bar color in phylogeny bars [ci skip]
