+++ R/graphs.R
@@ -107,8 +107,8 @@ abundance_heatmap_ggplot <- function(phyloseq_obj, classification = NULL,
       theme(
         axis.text.x = element_text(angle = 30, hjust = 1, size = 12),
         axis.text.y = element_text(hjust = 0.95, size = 12),
-        axis.title.x = element_blank(),
-        axis.title.y = element_blank(),
+        axis.title.x = element_text(size = 14, face = 'bold'),
+        axis.title.y = element_text(size = 14, face = 'bold'),
         axis.ticks.x = element_blank(),
         legend.title = element_text(size = 14, face = 'bold'),
         legend.text = element_text(size = 12),
