### 2019-06-12 42408b2
add CI ellipse options to ordinations
