+++ R/graphs.R
@@ -590,7 +590,7 @@ network_layout_ps <- function (phyloseq_obj, classification = NULL, treatment = 
 #' Inputs a \code{\link[phyloseq]{phyloseq-class}} object and
 #' plots the NMDS of a treatment or set of treatments in space.
 #' @useDynLib phylosmith
-#' @usage nmds_phyloseq_ggplot(phyloseq_obj, treatment, circle = TRUE,
+#' @usage nmds_phyloseq_ggplot(phyloseq_obj, treatment, circle = 0.95,
 #' labels = NULL, colors = 'default', verbose = TRUE)
 #' @param phyloseq_obj A \code{\link[phyloseq]{phyloseq-class}} object. It
 #' must contain \code{\link[phyloseq:sample_data]{sample_data()}}) with
@@ -600,8 +600,10 @@ network_layout_ps <- function (phyloseq_obj, classification = NULL, treatment = 
 #' @param treatment Column name as a string or number in the
 #' \code{\link[phyloseq:sample_data]{sample_data}}. This can be a vector of
 #' multiple columns and they will be combined into a new column.
-#' @param circle Add a \code{\link[ggplot2:stat_ellipse]{stat_ellipse}} around
-#' each of the \code{treatment} factors (\code{TRUE}).
+#' @param circle If TRUE, a \code{\link[ggplot2:stat_ellipse]{stat_ellipse}} around
+#' each of the \code{treatment} factors (\code{TRUE}). If numeric between 0 and 1,
+#' will add ellipse of confidence interval equal to value given (i.e. 0.95 produces
+#' ellipses of 95\% confidence intervals)
 #' @param labels Column name as a string or number in the
 #' \code{\link[phyloseq:sample_data]{sample_data}} to use to place labels of
 #' that factor instead of circle points.
@@ -618,7 +620,7 @@ network_layout_ps <- function (phyloseq_obj, classification = NULL, treatment = 
 #' @examples nmds_phyloseq_ggplot(soil_column, c('Matrix', 'Treatment'),
 #' circle = TRUE, verbose = FALSE)
 
-nmds_phyloseq_ggplot <- function(phyloseq_obj, treatment, circle = TRUE,
+nmds_phyloseq_ggplot <- function(phyloseq_obj, treatment, circle = 0.95,
     labels = NULL, colors = 'default', verbose = TRUE){
     if(!inherits(phyloseq_obj, "phyloseq")){
         stop("`phyloseq_obj` must be a phyloseq-class
@@ -633,9 +635,9 @@ nmds_phyloseq_ggplot <- function(phyloseq_obj, treatment, circle = TRUE,
         stop("`treatment` must be at least one column
         name, or index, from the sample_data()", call. = FALSE)
     }
-    if(!(is.logical(circle))){
-        stop("`circle` must be either `TRUE`, or
-        `FALSE`", call. = FALSE)
+    if(!(is.logical(circle) | is.numeric(circle) & circle > 0 & circle <= 1)){
+        stop("`circle` must be either `TRUE`, `FALSE`, or a confidence interval",
+             call. = FALSE)
     }
     labels <- check_numeric_treatment(phyloseq_obj, labels)
     if(!(is.null(labels)) & any(!(labels %in% colnames(access(phyloseq_obj,
@@ -650,27 +652,32 @@ nmds_phyloseq_ggplot <- function(phyloseq_obj, treatment, circle = TRUE,
     options(warn = -1)
     phyloseq_obj <- taxa_filter(phyloseq_obj, treatment, frequency = 0)
     treatment_name <- paste(treatment, collapse = sep)
-    Treatment <- access(phyloseq_obj, 'sam_data')[[treatment_name]]
-    color_count <- length(unique(Treatment))
[…]
