+++ man/nmds_phyloseq_ggplot.Rd
@@ -5,7 +5,7 @@
 \title{Create a ggplot object of the NMDS from a phyloseq object.
 Function from the phylosmith-package.}
 \usage{
-nmds_phyloseq_ggplot(phyloseq_obj, treatment, circle = TRUE,
+nmds_phyloseq_ggplot(phyloseq_obj, treatment, circle = 0.95,
 labels = NULL, colors = 'default', verbose = TRUE)
 }
 \arguments{
@@ -19,8 +19,10 @@ taxa/gene.}
 \code{\link[phyloseq:sample_data]{sample_data}}. This can be a vector of
 multiple columns and they will be combined into a new column.}
 
-\item{circle}{Add a \code{\link[ggplot2:stat_ellipse]{stat_ellipse}} around
-each of the \code{treatment} factors (\code{TRUE}).}
+\item{circle}{If TRUE, a \code{\link[ggplot2:stat_ellipse]{stat_ellipse}} around
+each of the \code{treatment} factors (\code{TRUE}). If numeric between 0 and 1,
+will add ellipse of confidence interval equal to value given (i.e. 0.95 produces
+ellipses of 95\% confidence intervals)}
 
 \item{labels}{Column name as a string or number in the
 \code{\link[phyloseq:sample_data]{sample_data}} to use to place labels of
