+++ man/CI_ellipse.Rd
@@ -1,0 +1,23 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/non_export.R
+\name{CI_ellipse}
+\alias{CI_ellipse}
+\title{Creates confidence interval ellipse for scatterplot}
+\usage{
+CI_ellipse(points, groups = NULL, level = 0.95)
+}
+\arguments{
+\item{points}{dataframe of the x,y coordinates of the datapoints, along with
+any metadata such as group}
+
+\item{groups}{Name of the column in points that has group membership}
+
+\item{level}{the confidence level to be plotted (default 95\%)}
+}
+\value{
+ellipse points
+}
+\description{
+Creates confidence interval ellipse for scatterplot. Can be created
+for each group or entire set.
+}
