+++ NAMESPACE
@@ -38,6 +38,8 @@ import(stringr)
 importFrom(Rtsne,Rtsne)
 importFrom(sf,st_as_sf)
 importFrom(sf,st_buffer)
+importFrom(stats,cov.wt)
+importFrom(stats,qf)
 importFrom(vegan,metaMDS)
 importFrom(vegan,scores)
 importFrom(vegan,vegdist)
