### 2019-06-16 #1442 @joey711
Also tested the `co_occurrence` function, which was re-implemented with Rcpp for performance. Can confirm the performance. I did not test accuracy.

Speaking of accuracy, that function in particular could use some unit tests to establish a baseline of analytical accuracy. A few other functional unit tests wouldn't hurt.
