### 2019-06-16 #1442 @joey711
Main deficiencies to address:

- (1) Installation. The documented installation code is insufficient. The phyloseq package is released through bioconductor, and is not installed automatically as a dependency through the indicated `devtools::install_github("phylosmith")` command. More importantly, there are other deeply-implied dependencies related to the Rcpp compilation. I was able to resolve this with the following `Dockerfile` code, starting with an official bioconductor release image (see below).
- (2) Unit tests. There are none.

Dockerfile code:

```
FROM bioconductor/release_core2

# Install phyloseq
RUN Rscript -e 'if(!requireNamespace("BiocManager", quietly = TRUE)){install.packages("BiocManager")}; BiocManager::install("phyloseq")'

RUN Rscript -e 'install.packages("devtools")'

RUN apt-get update && apt-get install -y --no-install-recommends libudunits2-dev
RUN Rscript -e 'install.packages("units")'

RUN Rscript -e 'install.packages("RcppEigen")'
RUN Rscript -e 'install.packages("RcppParallel")'
RUN Rscript -e 'install.packages("Rtsne")'

# ggforce
RUN apt-get install -y --no-install-recommends libgdal-dev
RUN Rscript -e 'install.packages("ggforce")'

# Install phylosmith
RUN Rscript -e 'devtools::install_github("schuyler-smith/phylosmith")'

# To run R from the command line:
# docker run -ti --user bioc bioconductor/release_core2 R
# docker run -e PASSWORD=rstudio1 -p 8787:8787 phylosmith
```
