### 2019-06-16 schuyler-smith/phylosmith#1 @joey711
Installation documentation insufficient
https://github.com/openjournals/joss-reviews/issues/1442

Here is an example Dockerfile that resolves the installation failures. There are multiple implicit dependencies not captured by the currently-provided installation code:

```
FROM bioconductor/release_core2

# Install phyloseq
RUN Rscript -e 'if(!requireNamespace("BiocManager", quietly = TRUE)){install.packages("BiocManager")}; BiocManager::install("phyloseq")'

RUN Rscript -e 'install.packages("devtools")'

RUN apt-get update && apt-get install -y --no-install-recommends libudunits2-dev
RUN Rscript -e 'install.packages("units")'

RUN Rscript -e 'install.packages("RcppEigen")'
RUN Rscript -e 'install.packages("RcppParallel")'
RUN Rscript -e 'install.packages("Rtsne")'

# ggforce
RUN apt-get install -y --no-install-recommends libgdal-dev
RUN Rscript -e 'install.packages("ggforce")'

# Install phylosmith
RUN Rscript -e 'devtools::install_github("schuyler-smith/phylosmith")'

# To run R from the command line:
# docker run -ti --user bioc bioconductor/release_core2 R
# docker run -e PASSWORD=rstudio1 -p 8787:8787 phylosmith
```

I'm providing this just as an example. The layers are disorganized due to debugging. Does serve as a MRE of the installation working from a phyloseq- and phylosmith-naive system state.
