### 2019-06-16 schuyler-smith/phylosmith#2 @joey711
Could use a few unit tests, especially `co_occurrence` function
`co_occurrence`

https://github.com/openjournals/joss-reviews/issues/1442
