### 2019-06-16 #1442 @joey711
Otherwise looks good. @karthik I'm not sure what to do next? I think author will find it is in their own best interest to address these two minor issues. Afterward, it seems to meet all the criteria and I'm happy to get out of the way.
