### 2019-06-17 eba984c
extend installation instructions
