+++ vignette_site/index.Rmd
@@ -19,12 +19,44 @@ A lot of these functions are just to make "data-wrangling" easier for the user. 
 <br>
 
 # Installation
-```r
-library(devtools)
-install_github('schuyler-smith/phylosmith')
+## Requirements
+### Linux Systems 
+For some Linux systems you may need to install the following two programs through your terminal.
+
+Ubuntu example:
+```
+sudo apt install libgdal-dev libudunits2-dev
+```
+These programs are required by some dependancies and may not come in your default OS distribution.
+
+### Windows Systems 
+
+if you are working on WINDOWS you likely need to install the CRAN program <a href="https://cran.r-project.org/bin/windows/Rtools/" target="_blank" >Rtools</a>.When prompted, select `add rtools to system PATH`.
+
+### R
+
+phylosmith depends on the usage of the [phyloseq package](https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0061217) released by Dr. Paul McMurdie. The pacakge is maintained on BioConductor, and can be installed through R using the following commands:
+
+```
+if(!requireNamespace("BiocManager", quietly = TRUE)){
+  install.packages("BiocManager")
+} 
+BiocManager::install("phyloseq")
+```
+
+Additionally, the package imports a number of other packages to use their advanced functions. These packages may install with the phylosmith installation, but it is always best to install independently.
+
+```
+install.packages(c("devtools", RcppEigen", "RcppParallel", "Rtsne", "ggforce", "units"))
+```
+
+## phylosmith
+
+The package is hosted on my Github page, and can be installed through R with:
+```
+devtools::install_github('schuyler-smith/phylosmith')
 library(phylosmith)
 ```
-\*for WINDOWS you need to install <a href="https://cran.r-project.org/bin/windows/Rtools/" target="_blank" >Rtools</a>, when prompted, select `add rtools to system PATH`.
 
 <br>
 
@@ -34,7 +66,7 @@ library(phylosmith)
 
 <a name="data_wrangling"></a>
 
-### [Data Wrangling][dapa]
+### Data Wrangling
 Call | Description
 -------------------- | ------------------------------------------------------------
 [common_taxa](data_wrangling.html#common_taxa) | find taxa common to each treatment
@@ -51,7 +83,7 @@ Call | Description
 
 <a name="graphs"></a>
 
-### [Graphs][grap]
+### Graphs
 Call | Description
 -------------------- | ------------------------------------------------------------
 [abundance_heatmap_ggplot](graphs.html#abundance_heatmap_ggplot) | create a ggplot object of the heatmaps of the abndance table
@@ -65,7 +97,7 @@ Call | Description
 
 <a name="calculations"></a>
 
-### [Calculations][calc]
+### Calculations
 Call | Description
 -------------------- | ------------------------------------------------------------
 [co_occurrence](calculations.html#co_occurrence) | calculate co-occurrence between taxa
