+++ vignettes/phylosmith.Rmd
@@ -17,18 +17,6 @@ A conglomeration of functions that I have written, that I find useful, for
 analyzing phyloseq objects. Phyloseq objects are a great data-standard for 
 microbiome and gene-expression data.
 
-# Installation
-
-```
-library(devtools)
-install_github('schuyler-smith/phylosmith')
-library(phylosmith)
-```
-
-*for WINDOWS you need to install <a href="https://cran.r-project.org/bin/
-windows/Rtools/" target="_blank" >Rtools</a>, when prompted, select `add 
-rtools to system PATH`.*
-
 # Functions
 
 ## [Data Wrangling](
