### 2019-06-17 a6b561e
extend installation instructions
