### 2019-06-17 #1442 @schuyler-smith
@joey711 Thank you for the review!
I was aware of the installation dependencies, from dealing with travis-ci, but wasn't sure how convoluted I should make the instructions vs let the user install dependencies on a need-to-need basis. I updated the instructions to include all the dependencies I included in the travis file, and the ones you listed for your installation. Hopefully it's clear and complete!

The unit tests I was admittedly not sure how to include, but had thought about. I am sure I have wasted a lot of my own time by not adding them sooner. I will look into including them!
