### 2019-06-17 #1442 @schuyler-smith
@karthik I addressed the installation documentation issues, hopefully to a satisfactory level. I will hope to add the unit tests in the near future. If they are needed, please let me know and I will make it a priority.

Is there anything I need to do for the next steps?
