### 2019-06-17 ebc99aa
add link to installation instructions
