### 2019-06-17 60b1ace
fix facets in heatmap
