### 2019-06-19 #1442 @schuyler-smith
@karthik also curious of what the next steps actually are, namely when doi's get assigned so that my labmate can cite. Thanks again for all the help!
