### 2019-06-19 #1442 @karthik
Thanks @joey711 for this review! @schuyler-smith I'm just returning from a trip so I'll need a day to look over the changes and recommend next steps. I'll update this thread shortly.
