### 2019-06-20 8bdc94f
replace sapply instances, update namespace
