+++ R/data_wrangling.R
@@ -26,53 +26,88 @@
 #' @examples common_taxa(soil_column, treatment = 'Treatment',
 #' subset = 'Amended', n = 'all')
 
-common_taxa <- function(phyloseq_obj, treatment = NULL, subset = NULL, n = 'all'){
-    if(!inherits(phyloseq_obj, "phyloseq")){
-        stop("`phyloseq_obj` must be a phyloseq-class object",
-            call. = FALSE)
+common_taxa <-
+  function(phyloseq_obj,
+           treatment = NULL,
+           subset = NULL,
+           n = 'all') {
+    if (!inherits(phyloseq_obj, "phyloseq")) {
+      stop("`phyloseq_obj` must be a phyloseq-class object",
+           call. = FALSE)
     }
-    if(is.null(access(phyloseq_obj, 'sam_data'))){
-        stop("`phyloseq_obj` must contain sample_data()
-        information", call. = FALSE)
+    if (is.null(access(phyloseq_obj, 'sam_data'))) {
+      stop("`phyloseq_obj` must contain sample_data()
+        information",
+           call. = FALSE)
     }
     treatment <- check_numeric_treatment(phyloseq_obj, treatment)
-    if(any(!(treatment %in% colnames(access(phyloseq_obj, 'sam_data'))))){
-        stop("`treatment` must be at least one column name, or
-        index, from the sample_data()", call. = FALSE)
+    if (any(!(treatment %in% colnames(access(
+      phyloseq_obj, 'sam_data'
+    ))))) {
+      stop(
+        "`treatment` must be at least one column name, or
+        index, from the sample_data()",
+        call. = FALSE
+      )
     }
-    if(!(is.numeric(n)) & n != 'all'){
-        stop("`n` must be either 'all' or a numeric value less
-        than the number of treatments being compared", call. = FALSE)
+    if (!(is.numeric(n)) & n != 'all') {
+      stop(
+        "`n` must be either 'all' or a numeric value less
+        than the number of treatments being compared",
+        call. = FALSE
+      )
     }
-    phyloseq_obj <- phyloseq(access(phyloseq_obj, 'otu_table'),
-                             access(phyloseq_obj, 'tax_table'), access(phyloseq_obj, 'sam_data'))
-    phyloseq_obj <- taxa_filter(phyloseq_obj, treatment, subset = subset)
+    phyloseq_obj <- phyloseq(
+      access(phyloseq_obj, 'otu_table'),
+      access(phyloseq_obj, 'tax_table'),
+      access(phyloseq_obj, 'sam_data')
+    )
+    phyloseq_obj <-
+      taxa_filter(phyloseq_obj, treatment, subset = subset)
     treatment_name <- paste(treatment, collapse = sep)
     treatment_classes <- unique(access(phyloseq_obj,
                                        'sam_data')[[treatment_name]])
-    if(n == 'all'){n <- length(treatment_classes)}
-    if(n > length(treatment_classes)){
-      stop("`n` must be either 'all' or a numeric value less
-        than the number of treatments being compared", call. = FALSE)
+    if (n == 'all') {
+      n <- length(treatment_classes)
     }
-    seen_taxa <- lapply(treatment_classes, FUN = function(trt){
-      taxa_names(taxa_filter(phyloseq_obj, treatment = treatment,
-                             subset = trt, frequency = 0))
-    })
-    if(n == 0){
-      seen_taxa <- tryCatch(taxa_names(taxa_filter(phyloseq_obj, frequency = .99)),
-                            error = function(e){stop("
+    if (n > length(treatment_classes)) {
+      stop(
+        "`n` must be either 'all' or a numeric value less
+        than the number of treatments being compared",
[…]
