+++ R/co_occurrence.R
@@ -20,11 +20,7 @@
 #' @param cores \code{numeric} Number of CPU cores to use for the pair-wise
 #' permutations. Default (0) uses max cores available. Parallelization not
 #' available for systems running MacOS without openMP configuration.
-#' @aliases FastCoOccur
-#' @import data.table
-#' @import RcppArmadillo
-#' @import RcppParallel
-#' @import RcppProgress
+#' @importFrom parallel detectCores
 #' @keywords nonparametric
 #' @seealso \code{\link{permute_rho}} \code{\link{phylosmith}}
 #' @export
@@ -35,58 +31,87 @@
 
 # sourceCpp("src/co_occurrence_Rcpp.cpp")
 
-co_occurrence <- function(phyloseq_obj, treatment = NULL, subset = NULL,
-                          rho = 0, p = 0.05, cores = 0){
-    if(!inherits(phyloseq_obj, "phyloseq")){
-        stop("`phyloseq_obj` must be a phyloseq-class object",
-        call. = FALSE)
+co_occurrence <-
+  function(phyloseq_obj,
+           treatment = NULL,
+           subset = NULL,
+           rho = 0,
+           p = 0.05,
+           cores = 0) {
+    if (!inherits(phyloseq_obj, "phyloseq")) {
+      stop("`phyloseq_obj` must be a phyloseq-class object",
+           call. = FALSE)
     }
     treatment <- check_numeric_treatment(phyloseq_obj, treatment)
-    if(!(is.null(treatment)) & is.null(access(phyloseq_obj, 'sam_data'))){
-        stop("`phyloseq_obj` must contain sample_data()
-        information if `treatment` argument is used", call. = FALSE)
+    if (!(is.null(treatment)) &
+        is.null(access(phyloseq_obj, 'sam_data'))) {
+      stop(
+        "`phyloseq_obj` must contain sample_data()
+        information if `treatment` argument is used",
+        call. = FALSE
+      )
     }
-    if(any(!(treatment %in% colnames(access(phyloseq_obj, 'sam_data'))))){
-        stop("`treatment` must be at least one column name,
-        or index, from the sample_data()", call. = FALSE)
+    if (any(!(treatment %in% colnames(access(
+      phyloseq_obj, 'sam_data'
+    ))))) {
+      stop(
+        "`treatment` must be at least one column name,
+        or index, from the sample_data()",
+        call. = FALSE
+      )
     }
-    if(!(is.numeric(p)) | !(p >= 0 & p <= 1)){
-        stop("`p` must be a numeric value between 0 and 1",
-        call. = FALSE)
+    if (!(is.numeric(p)) | !(p >= 0 & p <= 1)) {
+      stop("`p` must be a numeric value between 0 and 1",
+           call. = FALSE)
     }
-    if(!(is.numeric(rho)) | !(rho >= 0 & rho <= 1)){
-        stop("`rho` must be a numeric value between 0 and 1",
-        call. = FALSE)
+    if (!(is.numeric(rho)) | !(rho >= 0 & rho <= 1)) {
+      stop("`rho` must be a numeric value between 0 and 1",
+           call. = FALSE)
     }
-    if(!(is.numeric(cores)) |
-        !(cores >= 0 & cores <= (parallel::detectCores() -1))){
-        stop("`cores` must be a numeric value between 0 and ",
-        (parallel::detectCores() -1),
+    if (!(is.numeric(cores)) |
+        !(cores >= 0 & cores <= (detectCores() - 1))) {
+      stop(
+        "`cores` must be a numeric value between 0 and ",
+        (detectCores() - 1),
         "\n(upper limit is set by the cores available on machine used to
-        execute code)" , call. = FALSE)
[…]
