+++ man/processOBO.Rd
@@ -1,18 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/classify_ARGs.R
-\name{processOBO}
-\alias{processOBO}
-\title{processOBO}
-\usage{
-processOBO(OBO_filepath)
-}
-\arguments{
-\item{OBO_filepath}{filepath to the aro.obo file.}
-}
-\value{
-list
-}
-\description{
-Converts the aro.obo file from the
-\href{https://card.mcmaster.ca/home}{CARD database} into a data.table.
-}

[…]
