### 2019-06-20 d675132
correct spelling errors, update site
