+++ man/mock_phyloseq_2.Rd
@@ -14,7 +14,7 @@ data(mock_phyloseq_2)
 \description{
 These datasets were made-up for the purpose
 of demonstrating the examples of the functions.
-They are not biologicall accurate, and don't
+They are not biologically accurate, and don't
 represent anything meaningful beyond being
 functional phyloseq objects.
 }

[…]
