### 2019-06-20 6eb5420
remove card from sysdata
