### 2019-06-20 8789582
add zenodo doi, v1.0.0
