### 2019-06-20 #1442 @danielskatz
@schuyler-smith - Here are  bunch of corrections to the paper:

- "Presented here is phylosmith" could be "This paper presents phylosmith"  (and perhaps this will also fix the breaking phylosmith at the end of line 2 and the start of line 3)
- "Functions that will either return a transformed version of the input phyloseq object, a subset version, or extracted data from it on set parameters" is not a complete sentence.
- "which can suffers" should be "which can suffer"
- "from convergence getting stuck in local minima" could be "from converging to local minima"
- "As of this announcement," could be "As of publication of this paper,"
- "the age of big-data that biologists" could be "the age of big-data, biologists"
- "that their should" should be "that there should"
- "there has developed a formulaic approach to analysis" could be "a formulaic approach to analysis has developed"
- "i.e. incorrect" should be "e.g., incorrect"
- "and medium, users" should be "and intermediate users"
- "a barrier as they" should be "a barrier, as they"
- "the package data.table" could be "the data.table package"
- "on answering biological." needs another word at the end, perhaps "questions"
