### 2019-06-20 #1442 @danielskatz
@schuyler-smith - also, please update the title of the Zenodo archive to match the title of the paper: "phylosmith: an R-package for reproducible and efficient microbiome analysis with phyloseq-objects"
