+++ JOSS_paper/paper.md
@@ -24,34 +24,34 @@ output:
 
 # Summary
 
-Presented here is ``phylosmith``, an ``R``-package that enables reproducible and efficient analysis of microbiome data with ``phyloseq-class`` objects by providing robust and efficient functions. ``phylosmith`` utilizes the standardized data format of ``phyloseq`` and ``R`` object accession methods to provide functions with simple and intuitive input arguments. 
+This paper presents ``phylosmith``, an ``R``-package that enables reproducible and efficient analysis of microbiome data with ``phyloseq-class`` objects by providing robust and efficient functions. ``phylosmith`` utilizes the standardized data format of ``phyloseq`` and ``R`` object accession methods to provide functions with simple and intuitive input arguments. 
 
 The functions provided in ``phylosmith`` have been divided into 3 categories.
 
 ## Data Wrangling
 
-Functions that will either return a transformed version of the input ``phyloseq`` object, a subset version, or extracted data from it on set parameters. In some cases the functions are practical rewrites of ones already available from ``phyloseq`` with additional features or a more efficient implementation with ``data.table`` for large datasets. Various functions include finding shared or exclusive taxa by treatments, agglomeration, factor handling, and filtering.
+These functions include operations that will either return a transformed version of the input ``phyloseq`` object, a subset version, or extracted data on set parameters. In some cases the functions are practical rewrites of ones already available from ``phyloseq`` with additional features or a more efficient implementation with ``data.table`` for large datasets. Various functions include finding shared or exclusive taxa by treatments, agglomeration, factor handling, and filtering.
 
 ## Graphs
 
 The graphs are designed to serve as a quick and easy way to visualise data for 
 analysis and to provide a foundation for figures for publishing. Graphics include ordinations, phylogeny profiles, and co-occurrence networks. All images are produced as ``ggplot`` objects (@ggplot2), allowing for the image to be altered and additional layers given to tailor the graphic as desired. Additionally, the code for producing the graphs is readily accessible, allowing for the code to be reused and tailored to fit needs, providing a foundation to start from. The most novel, for the field of microbiome research, is the implementation of a t-SNE ordination. Most studies have 
-used PCA or NMDS, which can suffers from convergence getting stuck in local minima on large datasets, t-SNE is designed for large datasets and is not susceptible to these same limitations.
+used PCA or NMDS, which can suffer from converging to a local minima on large datasets, t-SNE is designed for large datasets and is not susceptible to these same limitations.
 
 ## Calculations
 
-As of this announcement, the functions in this section all pertain to calculating and analyzing the Spearman rank co-occurrence. The routine was written in efficient C++ code and interfaced with R using the ``Rcpp`` API (@Rcpp). The resulting co-occurrence table matches that produced by the ``cor()`` function in the ``R stats`` package, but is calculated much faster on a single thread, with a multi-threading options implemented as well.
+As of publication of this paper, the functions in this section all pertain to calculating and analyzing the Spearman rank co-occurrence. The routine was written in efficient C++ code and interfaced with R using the ``Rcpp`` API (@Rcpp). The resulting co-occurrence table matches that produced by the ``cor()`` function in the ``R stats`` package, but is calculated much faster on a single thread, with a multi-threading options implemented as well.
 
 # Need
 
-Adoption of data-standards enable data that are readily available for sharing and also the creation and implementation of tools for reproducible research. It is commonly said that in the age of big-data that biologists are required to have computational proficiency and literacy [@carey]. It seems reasonable that their should be a large onus on bioinformaticians to create accessible and practical tools that enable the biologists.
[…]
