### 2019-06-20 #1442 @schuyler-smith
@danielskatz thank you! This process has certainly opened my eyes to my inability to proofread my own work!

I addressed all the corrections you listed, and edited the Zenodo release title.
