### 2019-06-20 97e9b2e
bib fixes for JOSS
