+++ JOSS_paper/paper.bib
@@ -1,9 +1,9 @@
 
 @article{phyloseq,
-    author = {McMurdie, Paul J. AND Holmes, Susan},
+    author = {McMurdie, Paul J. and Holmes, Susan},
     journal = {PLOS ONE},
     publisher = {Public Library of Science},
-    title = {phyloseq: An R Package for Reproducible Interactive Analysis and Graphics of Microbiome Census Data},
+    title = {phyloseq: An {R} Package for Reproducible Interactive Analysis and Graphics of Microbiome Census Data},
     year = {2013},
     month = {04},
     volume = {8},
@@ -12,7 +12,7 @@
     doi = {10.1371/journal.pone.0061217}
 }
 
-@article{R,
+@book{R,
   title={R: A language and environment for statistical computing},
   author={R-Core and others},
   year={2013},
@@ -33,7 +33,7 @@
 }
 
 @article{Rcpp,
-  title={Rcpp: Seamless R and C++ integration},
+  title={Rcpp: Seamless {R} and {C}++ integration},
   author={Eddelbuettel, Dirk and Francois, Romain and Allaire, J and Ushey, Kevin and Kou, Qiang and Russel, N and Chambers, John and Bates, D},
   journal={Journal of Statistical Software},
   volume={40},
@@ -58,9 +58,9 @@
 }
 
 @article{tsne,
-  title={Visualizing data using t-SNE},
+  title={Visualizing data using t-{SNE}},
   author={Maaten, Laurens van der and Hinton, Geoffrey},
-  journal={Journal of machine learning research},
+  journal={Journal of Machine Learning Research},
   volume={9},
   number={Nov},
   pages={2579--2605},
@@ -70,7 +70,7 @@
 @article{generic1,
   title={Structure, function and diversity of the healthy human microbiome},
   author={Huttenhower, Curtis and Gevers, Dirk and Knight, Rob and Abubucker, Sahar and Badger, Jonathan H and Chinwalla, Asif T and Creasy, Heather H and Earl, Ashlee M and FitzGerald, Michael G and Fulton, Robert S and others},
-  journal={nature},
+  journal={Nature},
   volume={486},
   number={7402},
   pages={207},
@@ -82,7 +82,7 @@
 @article{generic2,
   title={A core gut microbiome in obese and lean twins},
   author={Turnbaugh, Peter J and Hamady, Micah and Yatsunenko, Tanya and Cantarel, Brandi L and Duncan, Alexis and Ley, Ruth E and Sogin, Mitchell L and Jones, William J and Roe, Bruce A and Affourtit, Jason P and others},
-  journal={nature},
+  journal={Nature},
   volume={457},
   number={7228},
   pages={480},
@@ -93,7 +93,7 @@
 @article{generic3,
   title={Enterotypes of the human gut microbiome},
   author={Arumugam, Manimozhiyan and Raes, Jeroen and Pelletier, Eric and Le Paslier, Denis and Yamada, Takuji and Mende, Daniel R and Fernandes, Gabriel R and Tap, Julien and Bruls, Thomas and Batto, Jean-Michel and others},
-  journal={nature},
+  journal={Nature},
   volume={473},
   number={7346},
   pages={174},
