### 2019-06-20 #1442 @danielskatz
@schuyler-smith - I also noticed some bib issues - suggested fixes in https://github.com/schuyler-smith/phylosmith/pull/4

In addition, some of the R & R package citations seem a little off - can you check the citations in the bib part of the pdf and make sure these look correct and complete?
