### 2019-06-20 #1442 @schuyler-smith
@danielskatz I merged your bib fixes. The data.table package was not officially published, that I can find. Matt Dowel just has extensive online presence for it, but I wanted to cite it. Maybe there is a better way?
ggplot2 appears to be cited most through Hadley Wickham's book, that I used for citation, so it doesn't have a DOI or anything of the sort. Any suggestion on that is welcome too.

Are these what you were referring too?
