+++ JOSS_paper/paper.bib
@@ -20,7 +20,7 @@
 }
 
 @article{datatable,
-  title={data. table: Extension of data. frame},
+  title={data.table: Extension of data.frame},
   author={Dowle, Matt and Short, T and Lianoglou, S and Saporta, R and Srinivasan, A and Antonyan, E},
   year={2014}
 }
