### 2019-06-20 #1442 @schuyler-smith
@danielskatz I see a couple corrections to make, sorry.
