### 2019-06-20 ad47ae1
fix R citation [ci skip]
