+++ JOSS_paper/paper.bib
@@ -12,11 +12,13 @@
     doi = {10.1371/journal.pone.0061217}
 }
 
-@book{R,
-  title={R: A language and environment for statistical computing},
-  author={R-Core and others},
-  year={2013},
-  publisher={Vienna, Austria}
+@Manual{,
+ title = {R: A Language and Environment for Statistical Computing},
+ author = {{R Core Team}},
+ organization = {R Foundation for Statistical Computing},
+ address = {Vienna, Austria},
+ year = {2014},
+ url = {http://www.R-project.org/},
 }
 
 @article{datatable,
