### 2019-06-20 #1442 @danielskatz
The first ref should have DOI https://doi.org/10.1038/nature09944 added
The fourth ref should have DOI http://dx.doi.org/10.18637/jss.v040.i08 added

there could also be more...
