### 2019-06-20 b0c06bd
fix R citation [ci skip]
