+++ JOSS_paper/paper.bib
@@ -12,7 +12,7 @@
     doi = {10.1371/journal.pone.0061217}
 }
 
-@Manual{,
+@Manual{R,
  title = {R: A Language and Environment for Statistical Computing},
  author = {{R Core Team}},
  organization = {R Foundation for Statistical Computing},
