### 2019-06-20 f130449
add DOIs
