+++ JOSS_paper/paper.bib
@@ -1,24 +1,24 @@
 
 @article{phyloseq,
-    author = {McMurdie, Paul J. and Holmes, Susan},
-    journal = {PLOS ONE},
-    publisher = {Public Library of Science},
-    title = {phyloseq: An {R} Package for Reproducible Interactive Analysis and Graphics of Microbiome Census Data},
-    year = {2013},
-    month = {04},
-    volume = {8},
-    pages = {1-11},
-    number = {4},
-    doi = {10.1371/journal.pone.0061217}
+  author={McMurdie, Paul J. and Holmes, Susan},
+  journal={PLOS ONE},
+  publisher={Public Library of Science},
+  title={phyloseq: An {R} Package for Reproducible Interactive Analysis and Graphics of Microbiome Census Data},
+  year={2013},
+  month={04},
+  volume={8},
+  pages={1-11},
+  number={4},
+  doi={10.1371/journal.pone.0061217}
 }
 
 @Manual{R,
- title = {R: A Language and Environment for Statistical Computing},
- author = {{R Core Team}},
- organization = {R Foundation for Statistical Computing},
- address = {Vienna, Austria},
- year = {2014},
- url = {http://www.R-project.org/},
+  title={R: A Language and Environment for Statistical Computing},
+  author={{R Core Team}},
+  organization={R Foundation for Statistical Computing},
+  address={Vienna, Austria},
+  year={2014},
+  url={http://www.R-project.org/}
 }
 
 @article{datatable,
@@ -41,22 +41,23 @@
   volume={40},
   number={8},
   pages={1--18},
-  year={2011}
+  year={2011},
+  doi={10.18637/jss.v040.i08}
 }
 
 @misc{carey,
-  author = {Carey, Maureen A. AND Papin, Jason A.},
-  journal = {PLOS Computational Biology},
-  publisher = {Public Library of Science},
-  title = {Ten simple rules for biologists learning to program},
-  year = {2018},
-  month = {01},
-  volume = {14},
-  url = {https://doi.org/10.1371/journal.pcbi.1005871},
-  pages = {1-11},
-  abstract = {},
-  number = {1},
-  doi = {10.1371/journal.pcbi.1005871}
+  author={Carey, Maureen A. AND Papin, Jason A.},
+  journal={PLOS Computational Biology},
+  publisher={Public Library of Science},
+  title={Ten simple rules for biologists learning to program},
+  year={2018},
+  month={01},
+  volume={14},
+  url={https://doi.org/10.1371/journal.pcbi.1005871},
+  pages={1-11},
+  abstract={},
+  number={1},
+  doi={10.1371/journal.pcbi.1005871}
 }
 
 @article{tsne,
@@ -78,7 +79,7 @@
   pages={207},
   year={2012},
   publisher={Nature Publishing Group},
-  doi = {10.1038/nature11234}
+  doi={10.1038/nature11234}
 }
 
 @article{generic2,
@@ -89,7 +90,8 @@
   number={7228},
   pages={480},
   year={2009},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/nature07540}
 }
 
 @article{generic3,
@@ -100,5 +102,6 @@
   number={7346},
   pages={174},
   year={2011},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/nature09944}
 }
