### 2019-06-20 #1442 @schuyler-smith
@danielskatz you are correct, I found a couple other DOIs. I think i blindly trusted the google scholar bib citations. Hopefully it looks better.
